requirements.txt
Flask
python-whois
dnspython
requests
